split up tabular comparison scenarios
