+++ __tests__/gui/tabular_comparison_1/all_details_wc.json
