+++ __tests__/gui/tabular_comparison_4/profile.json
@@ -1,0 +1,41 @@
+{
+  "id": 8,
+  "name": "Profile name",
+  "permission_type": "public",
+  "requires_authentication": false,
+  "geography_hierarchy": {
+    "id": 1,
+    "name": "2011 SA Boundaries",
+    "root_geography": {
+      "name": "South Africa Test",
+      "code": "ZA",
+      "level": "country",
+      "version": "2011 Boundaries"
+    },
+    "description": "",
+    "configuration": {
+      "default_version": "test",
+      "versions": [
+        "test"
+      ]
+    }
+  },
+  "description": "",
+  "configuration": {
+    "urls": [
+      "localhost"
+    ],
+    "default_panel": "point_data",
+    "tutorial": [],
+    "default_coordinates": {
+      "lat": -28.995409163308832,
+      "long": 25.093833387362697,
+      "zoom": 6
+    },
+    "point_markers": {
+      "clustering": {
+        "enabled": true
+      }
+    }
+  }
+}
