+++ __tests__/gui/tabular_comparison_4/profiles.json
@@ -1,0 +1,93 @@
+{
+   "count":1,
+   "next":null,
+   "previous":null,
+   "results":[
+      {
+         "id":8,
+         "name":"Youth Explorer",
+         "permission_type":"public",
+         "requires_authentication":false,
+         "geography_hierarchy":{
+            "id":1,
+            "name":"2011 SA Boundaries",
+            "root_geography":{
+               "name":"South Africa",
+               "code":"ZA",
+               "level":"country",
+               "version":"2011 Boundaries"
+            },
+            "description":""
+         },
+         "description":"",
+         "configuration":{
+            "urls":[
+               "beta.youthexplorer.org.za",
+               "geo.youthexplorer.org.za",
+               "localhost"
+            ],
+            "feedback":{
+               "url":"https://forms.gle/XeSZTxgpyA7G46J27",
+               "text":"Feedback",
+               "visible":true
+            },
+            "choropleth":{
+               "colors":[
+                  "#fef0d9",
+                  "#fdcc8a",
+                  "#fc8d59",
+                  "#e34a33",
+                  "#b30000"
+               ],
+               "opacity":0.7,
+               "opacity_over":0.8
+            },
+            "formatting":{
+               "decimal":",.2f",
+               "integer":",.3f",
+               "percentage":".1%"
+            },
+            "tile_layers":[
+               {
+                  "url":"https://{s}.basemaps.cartocdn.com/rastertiles/voyager_nolabels/{z}/{x}/{y}.png",
+                  "pane":"tilePane",
+                  "zIndex":200
+               },
+               {
+                  "url":"https://{s}.basemaps.cartocdn.com/rastertiles/voyager_only_labels/{z}/{x}/{y}.png",
+                  "pane":"labelsPanel",
+                  "zIndex":350
+               }
+            ],
+            "translations":{
+               "en":{
+                  "locations":"services",
+                  "Point Mapper":"Services Mapper"
+               }
+            },
+            "default_panel":"point_data",
+            "preferred_children":{
+               "country":[
+                  "province"
+               ],
+               "district":[
+                  "municipality",
+                  "mainplace",
+                  "ward"
+               ],
+               "province":[
+                  "district",
+                  "municipality"
+               ],
+               "mainplace":[
+                  "subplace"
+               ],
+               "municipality":[
+                  "mainplace",
+                  "ward"
+               ]
+            }
+         }
+      }
+   ]
+}
